Fixed variable size

+++ classes/convert/bitstream.cpp
@@ -50,7 +50,7 @@ quint32 BitStream::next()
 {
     quint32 result = 0;
     int i = this->mBlockSize - 1;
-    quint64 fill = this->mMatrix->options()->maskFill();
+    quint32 fill = this->mMatrix->options()->maskFill();
     while (i >= 0)
     {
         result = result << 1;
